Initial work on map
*Reversed outcome of empty string cases on City comparators, in order to
put empty strings at the end of the sorted list.
*Current problem: Pairing latitude and longitude with city names results
in an output where the names are sorted correctly, but all lats and lons
except the first show as "null, null". The first shows the lat and lon
of what should be the last entry in the list. With the comparator
change, this results in two empty strings.